Use a logwriter to print build logs
Fixes #109
